TVIST1-840: Updated hearing briefing logic
